requirements.txt
# Horizon Cinemas Booking System (HCBS)
# Python 3.12

# Backend (FastAPI)
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.1          # needed by PyMySQL for secure connections
pydantic[email]==2.9.2
python-dotenv==1.0.1

# Authentication
bcrypt==4.0.1
python-jose[cryptography]==3.3.0   # JWT encoding/decoding
passlib[bcrypt]==1.7.4

# Desktop GUI
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
httpx==0.27.2                 # HTTP client for GUI → API calls

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Utilities
python-dateutil==2.9.0